clip_service/requirements.txt
fastapi
uvicorn
transformers
torch
timm
einops
python-dotenv
Pillow
python-multipart
pydantic